+++ src/index.js
@@ -45,6 +45,7 @@ const contactStore = require('./models/contacts/contact-store');
 const tofuStore = require('./models/contacts/tofu-store');
 const chatStore = require('./models/chats/chat-store');
 const fileStore = require('./models/files/file-store');
+const volumeStore = require('./models/volumes/volume-store');
 const ghostStore = require('./models/mail/ghost-store');
 const mailStore = require('./models/mail/mail-store');
 const validation = require('./helpers/validation/field-validation');
@@ -63,6 +64,9 @@ const systemMessages = require('./helpers/system-messages');
 const serverSettings = require('./models/server-settings');
 const chatInviteStore = require('./models/chats/chat-invite-store');
 const Contact = require('./models/contacts/contact');
+const prombservable = require('./helpers/prombservable');
+const tracker = require('./models/update-tracker');
+
 
 // MEMO: Do NOT export NodeJsonStorage and NodeFileStream here for compatibility reasons
 module.exports = {
@@ -78,6 +82,7 @@ module.exports = {
     chatStore,
     chatInviteStore,
     fileStore,
+    volumeStore,
     ghostStore,
     mailStore,
     validation,
@@ -92,5 +97,9 @@ module.exports = {
     clientApp,
     systemMessages,
     serverSettings,
-    Contact // mostly for its utility static functions
+    Contact, // mostly for its utility static functions
+    prombservable,
+    __: {
+        tracker
+    }
 };